TraktCleaner/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.65.0  # For progress bars
loguru==0.7.0  # For advanced logging
